tests/unit/test_quotation_engine.py
from types import SimpleNamespace

import polars as pl
import pytest
from ldaca_web_app.api.workspaces.analyses.quotation_core import (
    compute_quote_dataframe,
    prepare_documents_payload,
)
from ldaca_web_app.core.services.quotation_client import (
    QuotationServiceError,
    extract_remote_quotations,
    normalise_engine_base_url,
)
from ldaca_web_app.models import QuotationEngineConfig, QuotationEngineType
from ldaca_web_app.settings import settings
from pydantic import AnyHttpUrl, TypeAdapter

HTTP_URL = TypeAdapter(AnyHttpUrl).validate_python


def test_engine_config_local_clears_url():
    cfg = QuotationEngineConfig(
        type=QuotationEngineType.LOCAL, url=HTTP_URL("http://example.com")
    )
    assert cfg.type is QuotationEngineType.LOCAL
    assert cfg.url is None


def test_engine_config_remote_requires_url():
    with pytest.raises(ValueError):
        QuotationEngineConfig(type=QuotationEngineType.REMOTE)


def test_normalise_engine_base_url_variants():
    assert (
        normalise_engine_base_url("http://localhost:8005")
        == "http://localhost:8005/api/v1/quotation"
    )
    assert (
        normalise_engine_base_url("http://localhost:8005/api/v1/quotation")
        == "http://localhost:8005/api/v1/quotation"
    )
    assert (
        normalise_engine_base_url("http://localhost:8005/api/v1/quotation/extract")
        == "http://localhost:8005/api/v1/quotation"
    )


@pytest.mark.asyncio
async def test_extract_remote_requires_remote_engine():
    cfg = QuotationEngineConfig()
    with pytest.raises(QuotationServiceError):
        await extract_remote_quotations(cfg, {})


def test_prepare_documents_payload_stable_order():
    df = pl.DataFrame({"text": ["a", "b", "c"]})
    docs = prepare_documents_payload(df, "text")
    assert list(docs.keys()) == ["0", "1", "2"]
    assert docs["0"]["text"] == "a"


@pytest.mark.asyncio
async def test_remote_compute_chunks_based_on_settings(monkeypatch):
    engine = QuotationEngineConfig(
        type=QuotationEngineType.REMOTE,
        url=HTTP_URL("http://engine"),
    )
    df = pl.DataFrame({"body": [f"doc-{i}" for i in range(5)]})
    node = SimpleNamespace(data=df)

    calls = []

    async def fake_extract(cfg, documents, *, options=None, timeout=None):
        calls.append(
            {
                "cfg": cfg,
                "documents": documents,
                "options": options,
                "timeout": timeout,
            }
        )
        return {
            "results": [
                {
                    "identifier": doc_id,
                    "quotes": [
                        {
                            "quote": f"quote-{doc_id}",
                            "quote_start_idx": 0,
                            "quote_end_idx": 1,
                        }
                    ],
                }
                for doc_id in documents.keys()
            ]
        }

    monkeypatch.setattr(settings, "quotation_service_max_batch_size", 2)

    result = await compute_quote_dataframe(
        node,
        df,
        "body",
        engine,
        extract_remote_fn=fake_extract,
        quotation_service_max_batch_size=settings.quotation_service_max_batch_size,
        quotation_service_timeout=settings.quotation_service_timeout,
    )

    assert len(calls) == 3  # 5 docs -> batches of 2,2,1
    assert [list(call["documents"].keys()) for call in calls] == [
        ["0", "1"],
        ["2", "3"],
        ["4"],
    ]
    assert result.columns == ["body", "quotation"]
    assert result.to_dicts() == [
        {
            "body": "doc-0",
            "quotation": [
                {
                    "speaker": None,
                    "speaker_start_idx": None,
                    "speaker_end_idx": None,
                    "quote": "quote-0",
                    "quote_start_idx": 0,
                    "quote_end_idx": 1,
                    "verb": None,
                    "verb_start_idx": None,
                    "verb_end_idx": None,
                    "quote_type": None,
                    "quote_token_count": None,
                    "is_floating_quote": None,
                    "quote_row_idx": 0,
                }
            ],
        },
        {
            "body": "doc-1",
            "quotation": [
                {
                    "speaker": None,
                    "speaker_start_idx": None,
                    "speaker_end_idx": None,
                    "quote": "quote-1",
                    "quote_start_idx": 0,
                    "quote_end_idx": 1,
                    "verb": None,
                    "verb_start_idx": None,
                    "verb_end_idx": None,
                    "quote_type": None,
                    "quote_token_count": None,
                    "is_floating_quote": None,
                    "quote_row_idx": 0,
                }
            ],
        },
        {
            "body": "doc-2",
            "quotation": [
                {
                    "speaker": None,
                    "speaker_start_idx": None,
                    "speaker_end_idx": None,
                    "quote": "quote-2",
                    "quote_start_idx": 0,
                    "quote_end_idx": 1,
                    "verb": None,
                    "verb_start_idx": None,
                    "verb_end_idx": None,
                    "quote_type": None,
                    "quote_token_count": None,
                    "is_floating_quote": None,
                    "quote_row_idx": 0,
                }
            ],
        },
        {
            "body": "doc-3",
            "quotation": [
                {
                    "speaker": None,
                    "speaker_start_idx": None,
                    "speaker_end_idx": None,
                    "quote": "quote-3",
                    "quote_start_idx": 0,
                    "quote_end_idx": 1,
                    "verb": None,
                    "verb_start_idx": None,
                    "verb_end_idx": None,
                    "quote_type": None,
                    "quote_token_count": None,
                    "is_floating_quote": None,
                    "quote_row_idx": 0,
                }
            ],
        },
        {
            "body": "doc-4",
            "quotation": [
                {
                    "speaker": None,
                    "speaker_start_idx": None,
                    "speaker_end_idx": None,
                    "quote": "quote-4",
                    "quote_start_idx": 0,
                    "quote_end_idx": 1,
                    "verb": None,
                    "verb_start_idx": None,
                    "verb_end_idx": None,
                    "quote_type": None,
                    "quote_token_count": None,
                    "is_floating_quote": None,
                    "quote_row_idx": 0,
                }
            ],
        },
    ]
