pyproject.toml
[project]
name = "ldaca-web-app"
version = "0.1.0"
description = "LDaCA Web App - FastAPI backend with bundled production frontend for the Language Data Commons of Australia"
readme = "README.md"
requires-python = ">=3.14"
license = { text = "MIT" }
authors = [{ name = "Australian Text Analytics Platform" }]
keywords = [
    "ldaca",
    "text-analytics",
    "corpus",
    "linguistics",
    "fastapi",
    "web-app",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering :: Information Analysis",
]
dependencies = [
    "aiosqlite>=0.21.0",
    "bertopic>=0.17.3",
    "fastapi-users[sqlalchemy]>=14.0.1",        # Includes fastapi, python-multipart, sqlalchemy
    "fastapi[standard]>=0.135.0",               # Includes uvicorn, pydantic, starlette
    "fastexcel>=0.18.0",
    "google-auth-oauthlib>=1.2.2",              # Includes google-auth
    "httpx>=0.27.0",
    "pydantic-settings>=2.0.0",
    "psutil>=6.1.0",                            # Process management for proper cleanup of child processes
    "docworkspace>=0.2.1",
    "ldaca-loader",
    "pyarrow>=16.0.0",
    "tbb>=2022.3.1 ; sys_platform != 'darwin'",
    "openai>=2.26.0",
    "pyairtable>=3.3.0",
]

[project.optional-dependencies]
deploy = ["jupyter-server-proxy>=4.4.0", "ipython>=9.5.0"]

[project.scripts]
ldaca-web-app = "ldaca_web_app.cli:main"
ldaca_web_app = "ldaca_web_app.cli:main"

[tool.pytest.ini_options]
addopts = ["-ra", "--strict-markers", "--strict-config"]
pythonpath = ["src"]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "workspace: marks tests related to workspace functionality",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = ["twine>=6.2.0", "pytest-asyncio>=0.23.0"]

[tool.hatch.build.targets.sdist]
include = ["/src", "/src/ldaca_web_app/resources/**"]

[tool.hatch.build.targets.wheel]
packages = ["src/ldaca_web_app"]
include = ["src/ldaca_web_app/resources/**"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[tool.uv.sources]
ldaca-loader = { git = "https://github.com/Australian-Text-Analytics-Platform/ldaca-tabulator", branch = "ldaca_web_app_integration" }
